requirements.txt
# ============================
# Google Generative AI (Updated versions)
# ============================
google-api-core==2.15.0 
protobuf==4.25.3
google-generativeai==0.8.4  # Updated version
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
httplib2==0.22.0
google-ai-generativelanguage==0.6.15
# ============================
# LangChain Ecosystem (Updated versions)
# ============================
langcodes==3.5.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-google-genai==2.0.6  # Keep this version
langchain-text-splitters==0.3.8
youtube-transcript-api==1.1.1

# ============================
# Chroma for Vector Store
# ============================
chromadb==1.0.15

# ============================
# Loaders for Unstructured Data
# ============================
unstructured==0.18.3
pytube==15.0.0
pypdf==5.7.0
markdown==3.7
openpyxl==3.1.5
python-pptx==1.0.2
lxml==5.4.0
pillow==11.1.0
xmltodict==0.14.2
beautifulsoup4==4.13.3
pandas==2.2.3

# ============================
# Embeddings & Utilities
# ============================
scikit-learn==1.6.1
libmagic==1.0
dataclasses-json==0.6.7
requests==2.32.3
numpy==1.26.4  # More stable than 2.x
tqdm==4.67.1
python-magic==0.4.27

# ============================
# Playwright & HTML Rendering
# ============================
playwright==1.54.0